refactor: Вынесла весь ui в отдельные кастомные views и настроила их отображение во viewControllers
